+++ src/components/modal/index.ts
@@ -1,0 +1,2 @@
+export { default as AddRestaurantModal } from "./AddRestaurantModal";
+export { default as RestaurantDetailModal } from "./RestaurantDetailModal";
